update removed qrReadFragment and appended to ClientAccessFragment
