--- app/src/main/java/com/puc/pi3_es_2024_t24/manager/ClientPhotoFragment.kt
@@ -1,33 +1,0 @@
-package com.puc.pi3_es_2024_t24.manager
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.navigation.fragment.findNavController
-import com.puc.pi3_es_2024_t24.R
-import com.puc.pi3_es_2024_t24.databinding.FragmentClientPhotoBinding
-import com.puc.pi3_es_2024_t24.databinding.FragmentQrCodeReadBinding
-import com.puc.pi3_es_2024_t24.utils.CameraFragmentDirections
-
-class ClientPhotoFragment : Fragment() {
-private lateinit var binding: FragmentClientPhotoBinding
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        binding = FragmentClientPhotoBinding.inflate(inflater, container, false)
-        binding.btn1access.setOnClickListener {
-            val action = ClientPhotoFragmentDirections.actionClientPhotoFragmentToCameraFragment(1)
-            findNavController().navigate(action)        }
-        binding.btn2access.setOnClickListener {
-            val action = ClientPhotoFragmentDirections.actionClientPhotoFragmentToCameraFragment(2)
-            findNavController().navigate(action)
-        }
-        return binding.root
-    }
-
-
-}
