+++ app/src/main/java/com/puc/pi3_es_2024_t24/utils/CameraFragment.kt
@@ -173,7 +173,7 @@ class CameraFragment : Fragment() {
     }
 
     private fun navigationQrCode(qr: String?) {
-        val action = CameraFragmentDirections.actionCameraFragmentToQrCodeReadFragment(qrCodeInfo = qr)
+        val action = CameraFragmentDirections.actionCameraFragmentToClientAccessFragment(qrCodeInfo = qr)
         findNavController().navigate(action)
     }
